New approach for the central Makefile used by PROW (#201)
PROW's central `Makefile` is changed to delegate commands (`make`
targets) to the `Makefiles` of individual go modules.
This approach requires a `Makefile` to be present in the directory of
all `go.mod` files.
This approach provides sensible defaults for the `make` targets that
PROW uses, and even for other common tasks, but also allows each go
module to provide its own unittest/lint/gosec/... implementation if
necessary.

This PR also re-enables `gosec` presubmit tests, since it fixes the
issue that was the reason for disabling them.