Pkg Specialization Pipeline tests (#178)
Add a "middle-sized test" that can test a chain/pipeline of multiple KRM
functions end-to-end, but doesn't need a Kubernetes cluster, because it
is built on the built-in Go testing framework.
This is less than a real integration or end-to-end test, but more than a
unit test, even more than the component tests of single KRM function.

Fixes https://github.com/nephio-project/nephio/issues/165