reworked bootstrap controllers (#212)
I had to rework the bootstrap controllers and split them in 2. bootstrap
packages and bootstrap secrets
the approach to only rely on secret does not reliably work on packages:
- a secret could be available but the package could be approved later in
which case the package never get installed. So changed this to rely on
packages for the bootstrap controller.

updated readme(s) to reflect the changes.