Remove replace in go.mod (#253)
In order for the automated docker build to work, we need to comment out
the replace line.

Merging this will trigger the docker build and push the image to our
registry.

Alternatively, instead of merging this we can any of these:
- Manually build and push
- Fix the docker build in the nephio-controller-manager to copy in the
entire controllers directory and tweak the replace statement
- Figure out something better than the way we do our multiple go.mod
files right now