Give docker builds a broader context, to fix including dependencies across go modules (#288)
It does what it says in the title.
It also removes some repetitions in our Makefiles, but it is intended to
be backward compatible in a sense that all Makefiles should have the
same targets as before.

This is a superset of @radoslawc 's former PR:
https://github.com/nephio-project/nephio/pull/286 , we synced on it and
agreed that this is the approach that we should go with.