configinject fn (#316)
this fn injects a configuration based on the dependency CR in the
package. it searches (by package name) for the resoruces identified in
the dependency CR in the adjacent packages. Once found it injects the
found resource in the current package.

it assumes if a reference is made the resource must be present.