Requeue in approval and generic specializer (#452)
In the most recent Porch builds, I have cleaned up a bunch of noisy
watch notifications. This is a good thing, except that our approval
controller and specializers were relying on the verbose watch events
being thrown out. This means that our e2e test times will go up a lot if
we bump Porch version. This PR resolves that, by requeueing when it
makes sense to. There is more we could do (for example watching
PackageVariants so we find out immediately when they flip to Ready
state), but this is a good start. Here are the timings I have:

| Test | R1 | Current Nephio<br/>New Porch| This PR Nephio<br/>New Porch
|
| ---- | ---- | ----------------------- | -------------------------|
| 001.sh | 235 secs | 540 secs | 235 secs |
| 002.sh | 730 secs | 1307 secs | 584 secs |
| 003.sh | 112 secs | 53 secs | 69 secs |
| 004.sh | 152 secs | 432 secs | 91 secs |
| 005.sh | 358 secs | 388 secs | 222 secs |
| 006.sh | 307 secs | 413 secs | 176 secs |
| 007.sh | 371 secs | 484 secs | 186 secs |
| 008.sh | 291 secs | 289 secs |67 secs |
| 009.sh | 26 secs | 30 secs | 31 secs |

I also added events in the approval controller for when it proposes and
approves the PackageRevision.