Refactor support for Mockery in the makefiles (#526)
The support for mockery is moved from all other makefiles and .mk files
into the `default-mockery.mk` file. Mockery related code in other
makefiles and .mk files is removed.

The included `default-go.mk` file now includes `default-mockery.mk`.

There are two targets in `default-mockery.mk`:

1. `install-mockery`: Installs mockery in docker or locally if docker is
not available
2. `generate-mocks`: Runs generation of the mocks for go interfaces

The targets above must be run explicitly.

The `generate-mocks` target looks for `.mockery.yaml` files in the repo
and it runs the mockery mock generator on each `.mockery.yaml` file it
finds. This has the nice effect of allowing `.mockery.yaml` files to be
in either the root of the repo or in subdirectories, so the choice of
placement of `.mockery.yaml` files is left to the developer.