Improve approval reconciler timings (#797)
Change approval controller PR Get to hit the api directly instead of
reading from local cache.
Adjust the reque duration to prevent race condition.

During debugging the approval delay issue reported
[here](https://github.com/nephio-project/nephio/issues/462) it became
apparent that the packagerev being fetched was a cached version which
didn't get updated for quite some time.
To circumvent this, we are retrieving the PR using the apiReader
interface which bypasses the local cache and hits the k8s api directly.