Add gen-kustomize-res (#841)
** _NOTE:_ ** 
Main review focus is on -
https://github.com/nephio-project/nephio/pull/841/files#diff-29fd056fb0053f017fa95b29a5cacf982bbece2204a6c735297f99e20db258c1

80%+ code coverage from tests.

README -
https://github.com/nephio-project/nephio/pull/841/files#diff-e1b41770c42c56af2ef1643ac198fe104b1e9917a426dc8783dd575fb20ca8c7

README Overview
...
Insert a Kustomize KRM, or if the resource already exists, update it's
resources list to include the paths of all non local-config KRM
resources. The function provides the ability to generate a
kustomization.yaml, containing the KRM to be applied to a cluster by a
gitops tool.

---------

Signed-off-by: Fiachra Corcoran <fiachra.corcoran@est.tech>