Nephio operator for focom to oran-provisioning requests (#849)
***

**__NOTE:__** All data under the /config directory is generated by the
operator sdk so does not require a full review.

***

Very basic implementation of FOCOM operator, which reconciles FOCOM
provisioning requests into oran provisioning requests. This PR contains
very basic functionality based on the POCs explored earlier, and with
the review, more functionality will be added if needed.

Reconciling FOCOM provisioning requests into oran provisioning requests,
will look for correct TemplateInfo referenced in the request. Current
validation only checks if the referenced TemplateInfo exists, with given
name and version. Template parameters are not validated yet, it will be
done in the future. Once the referenced template is validated, reconcile
will proceed to find OCloudId referenced in the request and make sure it
exists. From the OCloudId it will read the secret referenced in the
OCloud.

In this early stage, secret contains kubeconfig for the OCloud cluster,
which will be used to build client to interact with the remote cluster.
Using this client reconciler will produce Oran provisioning requests on
the remote cluster.

Later, this is planned to be changed to support REST-like interface on
the remote cluster, rather then using kubeconfig. Secret in this phase
should be endpoint and bearer token,

Project contains few test samples, using testenv, and there is usual e2e
test from the operator-sdk that can validate operator deployment. This
code is at the POC level, and needs more work to be production ready.

ORAN provisioning request CRD is located inside oran-provisioning-crd
folder. Focom CRDS are generated by operator-sdk/kubebuilder from go
structs in the api folder. Generated yaml files are located in
config/crd/bases.

Current code does not cater for more then one-shot operations, and
patching of ORAN provisioning requests is not supported yet. Validation
of the TemplateInfo parameters is not implemented yet, potentially this
can be done by validating webhook, making them immutable for a start.

---------

Signed-off-by: Fiachra Corcoran <fiachra.corcoran@est.tech>
Signed-off-by: Dejan Kitic <dejan.kitic@ericsson.com>
Co-authored-by: Fiachra Corcoran <fiachra.corcoran@est.tech>
Co-authored-by: Liam Fallon <35595825+liamfallon@users.noreply.github.com>