Fix generic specializer infinite loop (#967)
As part of the free5gc test 008, we copy, pull, mutate, push a UPF pkg.

https://github.com/nephio-project/test-infra/blob/main/e2e/tests/free5gc/008.sh#L53

The push fails due to the generic specializer continuously trying to
update the claims on the relevant resources.

The reconciler itself needs to be reworked properly to address the issue
and determine when it is "finished"
For now, this is a quick fix to get the test suite passing.

Signed-off-by: Fiachra Corcoran <fiachra.corcoran@est.tech>